requirements.txt
# Core Data Warehouse Dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
pandas==2.1.3
python-dotenv==1.0.0

# Data Extraction
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
lxml==4.9.3

# ETL & Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Data Quality & Validation
great-expectations==0.18.3

# Visualization & Dashboard
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.8.1
seaborn==0.13.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
python-decouple==3.8

# Utilities
tqdm==4.66.1
colorlog==6.7.0
click==8.1.7